Add control to enable/disable CORES On/OFF depend on HEAT. + tuning!
With new changes to Intelli Thermal code, we cant go higher than 75C
for All cores! as there is no cooldown C range. it's always try to have
max freq! so when above 75C it's jump too close to 88C the point of HOTBOOT!

With 75C Max we have stable device.