Replace Sepolicy binary with one from CM11. we can mount F2FS with seclabel.
After tests, to mount F2FS with Selinux in Enforce mode. we have to have support
for in in sepolicy binary!

https://android-review.googlesource.com/#/c/96593/1/fs_use

CM11 devs added that to binary.

@blastagator tested the binary with my kernel and stock based rom.
and it's works as should, no negative results.

root@g2:/ #mount | grep data
/dev/block/platform/msm_sdcc.1/by-name/userdata /data f2fs rw,seclabel,nosuid,nodev,noatime,nodiratime,background_gc=on,user_xattr,active_logs=6 0 0

this is the final solution for mounting F2FS to all partitions.

with enforce or permissive in Selinux.

Thanks to CM11 devs and to blastagator for this solution.