PROFILES: set ZEN as default I/O driver. works fast and stable.
Small tune for max temp in performance profile. may get to point of
overtemp in this condition:

charger connected! GPS is on, GPS guide app is running. BT transmitting music!

all that create massive heat. and in some point kernel will hit the alarm!
and shutdown.

few degres fix that.